feat(major): eslint 3, stop relying on airbnb and too much invisible defaults
The default setup for our eslint-config-algolia was starting to be too
dependent on a lot of other packages which made us in a situation where
we could not upgrade easily.

I have created a complete config with some new defaults given eslint 3.

If you are using npm 3, the only package to install is
eslint-config-algolia.

Then since eslint-config-algolia depends on eslint, eslint-plugin-react
and eslint-plugin-import, and npm 3 flattens by default, it should work.